Add files via upload
Program uses discreet, event based simulation to model the queuing and service of a set of requests.
Input consists of the following data:
 The number of primary servers in the system.
 The number of secondary servers in the system.
 A set of service requests each consisting of an arrival time and two service
times in the order primary followed by secondary. This set is terminated by a
dummy record with arrival time and service times all equal to 0. (Note: the
arrival times are sorted in ascending order).

The simulation is to be of a system with two sets of servers, primary and secondary,
each set of servers associated with a queue. Customers arrive in the system and are
served first by a primary server and, on completion of this service, by a secondary
server. If all servers of a particular type are busy, the customer will enter either the
primary or secondary queue as appropriate.